refactor(handlers): move session trace context wiring to util
